03_ai_response_testing/test_llm_response.py
"""
Phase 3 - AI Response Testing
Test an LLM API response for quality, latency, and consistency.
This is where your Performance Testing + QA skills meet AI!
"""
import time
import os
from dotenv import load_dotenv
from rich.console import Console

load_dotenv()
console = Console()

def simulate_llm_response_test():
    """
    Simulates calling an LLM API and validating the response.
    Replace the simulated_response with actual OpenAI API call when ready.
    """
    TEST_PROMPT = "What is the difference between load testing and stress testing? Answer in 2 sentences."

    # Simulated response (replace with actual openai.chat.completions.create call)
    simulated_response = {
        "content": "Load testing evaluates system behavior under expected load conditions, measuring performance metrics like response time and throughput. Stress testing pushes the system beyond its limits to find the breaking point and observe failure behavior.",
        "tokens_used": 48,
        "model": "gpt-3.5-turbo (simulated)"
    }

    start_time = time.time()
    time.sleep(0.3)  # Simulated API latency
    end_time = time.time()
    latency_ms = round((end_time - start_time) * 1000, 2)

    # ✅ Validation Checkpoints (your QA instincts!)
    validations = {
        "Response not empty": len(simulated_response["content"]) > 0,
        "Response under 500 chars": len(simulated_response["content"]) < 500,
        "Latency under 2000ms": latency_ms < 2000,
        "Contains keyword 'load'": "load" in simulated_response["content"].lower(),
        "Contains keyword 'stress'": "stress" in simulated_response["content"].lower(),
        "Token count reasonable": simulated_response["tokens_used"] < 200,
    }

    console.print(f"\n[bold cyan]🤖 LLM Response Validation Report[/bold cyan]")
    console.print(f"Prompt   : {TEST_PROMPT}")
    console.print(f"Model    : {simulated_response['model']}")
    console.print(f"Latency  : {latency_ms} ms")
    console.print(f"Tokens   : {simulated_response['tokens_used']}")
    console.print(f"\n[bold]Response:[/bold]\n{simulated_response['content']}")
    console.print("\n[bold]✅ Validation Results:[/bold]")

    all_pass = True
    for check, passed in validations.items():
        status = "[green]PASS[/green]" if passed else "[red]FAIL[/red]"
        console.print(f"  {status}  {check}")
        if not passed:
            all_pass = False

    console.print(f"\n{'[bold green]All checks passed! ✅' if all_pass else '[bold red]Some checks failed! ❌'}")

if __name__ == "__main__":
    simulate_llm_response_test()
