requirements.txt
# Core HTTP & Load Testing
requests==2.31.0
aiohttp==3.9.1
locust==2.20.0

# AI / LLM APIs
openai==1.12.0
google-generativeai==0.4.0

# Data & Metrics
pandas==2.2.0
matplotlib==3.8.2
rich==13.7.0

# Testing Framework
pytest==8.0.0
pytest-asyncio==0.23.4

# Environment
python-dotenv==1.0.1
